Phase 5: Remove legacy SpeedyPhysics class + retire conftest module-clearing
Per PR #429 review: with composable physics in place, the parallel
SpeedyPhysics class duplicates all the same logic. Major version, BC
break is acceptable.

What changed
------------

- Deleted SpeedyPhysics class from jcm/physics/speedy/speedy_physics.py.
  Inlined the standalone set_physics_flags helper into speedy_terms.py
  (only consumer) and removed the now-empty file.
- Deleted speedy_physics_test.py (single integration test, fully covered
  by composable speedy_terms_test.py).
- model.py: default physics is now speedy_physics() factory.
- model_test.py: switched from SpeedyPhysics(...) to speedy_physics(...).
- Removed legacy-vs-composable equivalence test from
  composable_physics_slow_test.py.

Conftest cleanup
----------------

Removed conftest.py's autouse fixture that wiped jcm.* from sys.modules
between tests. With nnx.Module + tree_math.struct types being globally
registered, that wipe left stale pytree class identities in JAX's
compile cache; nnx.grad on composable physics then failed with "different
tree structures" (same name, different identity) errors that did not
reproduce in isolation.

Without the wipe, the full fast suite passes cleanly (652 tests). The
fixture was added years ago to defend against module-level state in the
old monolithic physics; that justification is gone now.
